requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.16.1
frozenlist==1.5.0
h11==0.14.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
numpy==1.24.4
outcome==1.3.0.post0
packaging==24.2
pillow==10.4.0
propcache==0.2.0
pyarrow==17.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
selenium==4.27.1
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.44.0
tabulate==0.9.0
together==1.3.14
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.12.2
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.15.2
